+++ packages/aws-sqs/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @mondrian-framework/aws-sqs
 
+## 1.0.7
+
+### Patch Changes
+
+- Update context creation
+- Updated dependencies
+  - @mondrian-framework/module@1.0.7
+  - @mondrian-framework/model@1.0.7
+
 ## 1.0.6
 
 ### Patch Changes