+++ packages/graphql-yoga/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @mondrian-framework/graphql-yoga
 
+## 1.0.13
+
+### Patch Changes
+
+- Update context creation
+- Updated dependencies
+  - @mondrian-framework/graphql@1.0.13
+  - @mondrian-framework/module@1.0.7
+  - @mondrian-framework/model@1.0.7
+
 ## 1.0.12
 
 ### Patch Changes